+++ Data/BoardApi.cs
@@ -11,7 +11,7 @@ namespace Data
 
         public abstract List<BallApi> GetAllBalls();
         public abstract void RemoveAllBalls();
-        public abstract BallApi AddBall(float xPosition, float yPosition, int radius, int weight, int xSpeed = 0, int ySpeed = 0);
+        public abstract BallApi AddBall(float xPosition, float yPosition, int radius, float weight, float xSpeed = 0, float ySpeed = 0);
 
         public static BoardApi CreateApi(int boardWidth, int boardHeight)
         {
